+++ ms18-database/teamC/Context/Tables/Member.cs
