+++ ms18-database/teamC/Context/Tables/MemberPermission.cs
