+++ ms18-database/teamC/Context/Tables/Photo.cs
