+++ ms18-database/teamC/Context/Tables/Receipt.cs
