+++ ms18-database/teamC/Context/Tables/ReceiptApproval.cs
