+++ ms18-database/teamC/Context/Tables/ReceiptStatus.cs
