+++ ms18-database/teamC/Context/Tables/Store.cs
