+++ ms18-database/teamC/Interfaces/IMemberRepository.cs
