+++ ms18-database/teamC/Interfaces/IPhotoRepository.cs
