+++ ms18-database/teamC/Interfaces/IReceiptRepository.cs
